latitude & longitude in AgentInfo: long -> double

+++ src/edu/fudan/se/bean/AgentOffer.java
@@ -4,8 +4,8 @@ public class AgentOffer extends AgentInfo {
 
 	public final int offer;
 	
-	public AgentOffer(String guid, String capacity, long longitude,
-			long latitude, double reputation, int offer) {
+	public AgentOffer(String guid, String capacity, double longitude,
+					  double latitude, double reputation, int offer) {
 		super(guid, capacity, longitude, latitude, reputation);
 		this.offer = offer;
 	}
